+++ src/Blocks/Callout.php
@@ -15,8 +15,8 @@ use Notion\Common\RichText;
  *
  * @psalm-type CalloutJson = array{
  *      callout: array{
- *          text: RichTextJson[],
- *          children: array{ type: BlockJson },
+ *          text: list<RichTextJson>,
+ *          children: list<BlockJson>,
  *          icon: EmojiJson|FileJson,
  *      },
  * }
@@ -27,17 +27,17 @@ class Callout implements BlockInterface
 
     private Block $block;
 
-    /** @var \Notion\Common\RichText[] */
+    /** @var list<RichText> */
     private array $text;
 
     private Emoji|File $icon;
 
-    /** @var \Notion\Blocks\BlockInterface[] */
+    /** @var list<\Notion\Blocks\BlockInterface> */
     private array $children;
 
     /**
-     * @param \Notion\Common\RichText[] $text
-     * @param \Notion\Blocks\BlockInterface[] $children
+     * @param list<RichText> $text
+     * @param list<\Notion\Blocks\BlockInterface> $children
      */
     private function __construct(
         Block $block,
@@ -139,7 +139,8 @@ class Callout implements BlockInterface
         return $this->children;
     }
 
-    public function withText(RichText ...$text): self
+    /** @param list<RichText> $text */
+    public function withText(array $text): self
     {
         return new self($this->block, $text, $this->icon, $this->children);
     }
@@ -157,7 +158,8 @@ class Callout implements BlockInterface
         return new self($this->block, $this->text, $icon, $this->children);
     }
 
-    public function withChildren(BlockInterface ...$children): self
+    /** @param list<BlockInterface> $children */
+    public function withChildren(array $children): self
     {
         $hasChildren = (count($children) > 0);