+++ tests/Unit/Blocks/BookmarkTest.php
@@ -83,7 +83,7 @@ class BookmarkTest extends TestCase
     public function test_replace_caption(): void
     {
         $caption = [ RichText::createText("Bookmark caption") ];
-        $bookmark = Bookmark::create("https://my-site.com")->withCaption(...$caption);
+        $bookmark = Bookmark::create("https://my-site.com")->withCaption($caption);
 
         $this->assertEquals($caption, $bookmark->caption());
     }